move auth session helper functions out of account
